+++ integration-tests/emulated/tests/people/people-polkadot/src/tests/reap_identity.rs
@@ -301,8 +301,7 @@ fn assert_reap_id_relay(total_deposit: Balance, id: &Identity) {
 
 		assert_ok!(PolkadotIdentityMigrator::reap_identity(
 			// Note: Root for launch testing, Signed once we open migrations.
-			// PolkadotOrigin::signed(PolkadotRelaySender::get()),
-			PolkadotOrigin::root(),
+			PolkadotOrigin::signed(PolkadotRelaySender::get()),
 			PolkadotRelaySender::get()
 		));