+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog for the runtimes governed by the Polkadot Fellowship.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.7] 14.06.2024
+
+Note: This release only affects the following runtimes and is not a full system release:
+
+- Polkadot Relay Chain
+- Polkadot People
+
+### Changed
+
+- Updated Relay and People configurations to complete launch ([polkadot-fellows/runtimes#350](https://github.com/polkadot-fellows/runtimes/pull/350))
+
 ## [1.2.6] 13.06.2024
 
 Note: This release only affects the following runtimes and is not a full system release: